% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{cfs2cms}
\alias{cfs2cms}
\title{Converts values from Cubic Feet per second to Cubic meters per second}
\usage{
cfs2cms(cfs, messages = FALSE)
}
\arguments{
\item{cfs}{numeric value for cubic feet per second}

\item{messages}{Boolean TRUE or FALSE; default = FALSE, use messages = T to get more information}
}
\value{
numeric value for cubic meters per second
}
\description{
Converts values from Cubic Feet per second to Cubic meters per second
}
\examples{
cfs2cms(cfs = 8000, messages = TRUE)
cfs2cms(cfs = 10000, messages = FALSE)
}
