% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{cms2cfs}
\alias{cms2cfs}
\title{Converts values from Cubic Meters per second to Cubic feet per second}
\usage{
cms2cfs(cms, messages = FALSE)
}
\arguments{
\item{cms}{numeric value for cubic meters per second}

\item{messages}{Boolean TRUE or FALSE; default = FALSE, use messages = T to get more information}
}
\value{
numeric value for cubic feet per second
}
\description{
Converts values from Cubic Meters per second to Cubic feet per second
}
\examples{
cms2cfs(cms = 227, messages = TRUE)
cms2cfs(cms = 445, messages = FALSE)
}
