% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{convert_to_xts}
\alias{convert_to_xts}
\title{Turn Formatted discharge data into an xts time-series object for plotting with dygraph}
\usage{
convert_to_xts(data, column_choice = "Discharge_cfs")
}
\arguments{
\item{formattedQ}{data.frame or tibble. formatted data (or subset and formatted data) output from convertQdataRaw2FormattedQ()}

\item{unit}{string. default unit = 'cfs', can also be unit = 'cms'}
}
\value{
xts time series object
}
\description{
Turn Formatted discharge data into an xts time-series object for plotting with dygraph
}
\examples{
Qdat_xts_cfs <- convert_to_xts(SampleQ_1week_formatted)
Qdat_xts_cms <- convert_to_xts(SampleQ_1week_formatted, column_choice ="Discharge_cms")
}
