% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{convertQdataRaw2FormattedQ}
\alias{convertQdataRaw2FormattedQ}
\title{Process NWIS gage data, remove unwanted columns, create CFS column, change column names}
\usage{
convertQdataRaw2FormattedQ(Qdata, xcol_index = 4, DTcol_index = 3)
}
\arguments{
\item{Qdata}{data.frame or tibble. from NWIS instantaneous values for discharge (Q)}

\item{xcol_index}{numeric. default = 4; Provide the index for the column containing discharge (cfs)}

\item{DTcol_index}{numeric. default = 3; Provide the index for the column containing time stamp or datetime}
}
\value{
xts object
}
\description{
Process NWIS gage data, remove unwanted columns, create CFS column, change column names
}
\examples{
Qdat_formatted <- convertQdataRaw2FormattedQ(SampleQ_1week_raw)

}
