% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageDischargeFunctions.R
\name{find_Q_from_WSE}
\alias{find_Q_from_WSE}
\title{Find Discharge for a Given Stage Elevation}
\usage{
find_Q_from_WSE(Sitename, Grid, Elevation)
}
\arguments{
\item{Sitename}{The name of the site.}

\item{Grid}{A data frame or list containing the stage-discharge grid.}

\item{Elevation}{The elevation for which to find the corresponding discharge.}
}
\value{
A data frame row or similar structure containing the discharge information at the specified elevation.
}
\description{
Determines the discharge at a specified stage elevation using a provided stage-discharge grid.
}
\examples{
\dontrun{
  discharge_info <- find_Q_from_WSE('2201R', grid, 400.002) # Assumes '2201R' is a valid site and grid is defined
}
}
