% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLagtime.R
\name{findLagtime}
\alias{findLagtime}
\title{Calculate Lag Time Based on River Mile}
\usage{
findLagtime(rivermile, message = F)
}
\arguments{
\item{rivermile}{Numeric; the river mile for which the lag time is to be calculated.}

\item{message}{Logical; if \code{TRUE}, prints the details of the calculation including the selected river mile,
nearest gage name, distance to the gage, and calculated lag hours. Defaults to \code{FALSE}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{lagtime}: A \code{lubridate::duration} object representing the estimated lag time.
\item \code{nearest_gage}: The name of the nearest upstream gage used for the calculation.
\item \code{nearest_gage_index}: The index of the nearest gage within the \code{GageList} dataframe.
}
}
\description{
This function estimates the lag time for water to travel from a given river mile to a specified point
based on the nearest upstream gage. It uses a linear model with coefficients (slope and intercept) provided
in a global \code{GageList} dataframe to compute the lag time in hours. The function allows for optional printing
of intermediate results, including the selected river mile, nearest gage name, distance to the gage in river
miles, and the calculated lag hours.
}
\examples{
\dontrun{
  out <- findLagtime(rivermile = 32.9, message = TRUE)
  print(out)
}

}
