% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findQfromWSE.R
\name{findQfromWSE}
\alias{findQfromWSE}
\title{Find Median Predicted Discharge from Water Surface Elevation}
\usage{
findQfromWSE(Sitename, Elevation, Unit = "CFS")
}
\arguments{
\item{Sitename}{A string representing the site name; must be a valid site within the preloaded \code{SandbarSites} data.}

\item{Elevation}{A numeric value representing the water surface elevation in meters; must be within the valid range for the site.}

\item{Unit}{The unit of discharge: either "CFS" (Cubic Feet per Second) or "CMS" (Cubic Meters per Second).
Default is "CFS".}
}
\value{
A list containing discharge information at the given elevation, including lower, median, and upper estimates in the specified unit.
Access specific discharge information using the keys from the list like \code{ReturnedList$CFS_median} or \code{ReturnedList$CMS_median}.
}
\description{
This function calculates the median predicted discharge for a given water surface elevation
at a specified site using precomputed stage-discharge relationships. It checks for valid site name,
elevation range, and unit of discharge before proceeding.
}
\details{
The function first verifies the site name and elevation against pre-stored values.
If the site name or elevation is invalid, it throws an error.
It also checks if the provided unit is valid and throws an error if not.
After validation, it retrieves the closest grid row to the specified elevation from a precomputed dataframe,
converts the relevant row to a list, and returns the discharge values for the median prediction.
}
\examples{
out <- findQfromWSE("0081L", 915.345)

}
