% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageDischargeFunctions.R
\name{findSiteElevation}
\alias{findSiteElevation}
\title{Find Site Elevation Range}
\usage{
findSiteElevation(sitename)
}
\arguments{
\item{sitename}{A character string specifying the name of the sandbar site.}
}
\value{
A numeric vector with two elements: the minimum and maximum elevation.
}
\description{
Determines the minimum and maximum stage (elevation) recorded at a specified sandbar site.
}
\examples{
elevation_range <- findSiteElevation('2023R') # Assumes '2023R' is a valid site in SandbarSites

}
