% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageDischargeFunctions.R
\name{fitGeneralizedPowerLawModel}
\alias{fitGeneralizedPowerLawModel}
\title{Fit Generalized Power Law Model}
\usage{
fitGeneralizedPowerLawModel(sitename)
}
\arguments{
\item{sitename}{A character string specifying the name of the sandbar site.}
}
\value{
An object of class \code{gplm} containing the fitted model.
}
\description{
Fits a generalized power law model to stage-discharge data for a specified sandbar site using the \code{bdrc} package.
}
\examples{
\dontrun{
  model <- fitGeneralizedPowerLawModel('0307R') # Assumes '0307R' is a valid site in SandbarSites
}
}
