% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{FormattedQ2xts}
\alias{FormattedQ2xts}
\title{Turn Formatted discharge data into an xts time-series object for plotting with dygraph}
\usage{
FormattedQ2xts(formattedQ, unit = "cfs")
}
\arguments{
\item{formattedQ}{data.frame or tibble. formatted data (or subset and formatted data) output from convertQdataRaw2FormattedQ()}

\item{unit}{string. default unit = 'cfs', can also be unit = 'cms'}
}
\value{
xts time series object
}
\description{
Turn Formatted discharge data into an xts time-series object for plotting with dygraph
}
\examples{
Qdat_xts_cfs <- FormattedQ2xts(SampleQ_1week_formatted)
Qdat_xts_cms <- FormattedQ2xts(SampleQ_1week_formatted, unit='cms')
}
