% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageDischargeFunctions.R
\name{GenerateEquallySpacedStageDischarge}
\alias{GenerateEquallySpacedStageDischarge}
\title{Generate Equally Spaced Stage-Discharge Grid}
\usage{
GenerateEquallySpacedStageDischarge(sitename, ElevIncrement = 0.001)
}
\arguments{
\item{sitename}{A character string specifying the name of the sandbar site.}

\item{ElevIncrement}{The elevation increment for the grid.}
}
\value{
A data frame containing the elevation grid and corresponding discharge predictions.
}
\description{
Generates a grid of equally spaced stage values and their corresponding discharge predictions from a fitted model.
}
\examples{
\dontrun{
  grid <- GenerateEquallySpacedStageDischarge('2201R')
}
}
