% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestUpstreamGage.R
\name{getNearestUpstreamGage}
\alias{getNearestUpstreamGage}
\title{function to find the nearest upstream gage in Grand Canyon to the desired rivermile}
\usage{
getNearestUpstreamGage(riverMile, messages = F)
}
\arguments{
\item{riverMile}{numeric value between 0 and 283}

\item{messages}{Boolean T or F, default FALSE; TRUE shows messages, FALSE suppresses them}
}
\value{
int the index for the closest upstream gage in the dataframe (tibble) GageList
}
\description{
function to find the nearest upstream gage in Grand Canyon to the desired rivermile
}
\examples{
index <- getNearestUpstreamGage(23)
index <- getNearestUpstreamGage(103, TRUE)
}
