% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHydrograph.R
\name{plotFlowDurationCurve}
\alias{plotFlowDurationCurve}
\title{Plot a Flow Duration Curve}
\usage{
plotFlowDurationCurve(Qdat, unit = "CFS", title = "Flow Duration Curve")
}
\arguments{
\item{Qdat}{A dataframe containing the discharge data to be plotted, expected to include
a POSIXct dateTime column, and numeric Discharge_cfs and Discharge_cms columns.}

\item{unit}{A character string specifying the unit of the discharge data. Supported
units are 'CFS' for cubic feet per second and 'CMS' for cubic meters per second.
Default is 'CFS'.}

\item{title}{Optional: The main title of the flow duration curve. Default is "Flow Duration Curve".}
}
\value{
A ggplot object representing the flow duration curve.
}
\description{
Creates a flow duration curve using the \code{ggplot2} package. This function is designed to plot
flow duration data from a given discharge dataset in either cubic feet per second (CFS)
or cubic meters per second (CMS). The curve represents the percentage of the time that certain
flows are equaled or exceeded during the period of record.
}
\examples{
plotFlowDurationCurve(Qdat = SampleQ_1week_processed, unit = 'CMS', title = "Flow Duration Curve for Colorado River")

}
