% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHydrograph.R
\name{plotInteractiveHydrograph}
\alias{plotInteractiveHydrograph}
\title{Plot an Interactive Hydrograph}
\usage{
plotInteractiveHydrograph(
  Qdat,
  title = "Hydrograph",
  subtitle = "subtitle",
  unit = "cfs"
)
}
\arguments{
\item{Qdat}{A time series object (preferably an xts object) containing the discharge
data to be plotted.}

\item{title}{The main title of the hydrograph plot. Default is "Hydrograph".}

\item{subtitle}{The subtitle of the hydrograph plot, typically used for specifying
the date range or other descriptive information. Default is "subtitle".}

\item{unit}{A character string specifying the unit of the discharge data. Supported
units are 'cfs' for cubic feet per second and 'cms' for cubic meters per second.
Default is 'cfs'.}
}
\value{
An interactive hydrograph plot rendered using \code{dygraphs}.
}
\description{
Creates an interactive hydrograph using the \code{dygraphs} package. This function
is designed to plot time series data of water discharge, allowing for the
specification of units in cubic feet per second (cfs) or cubic meters per second
(cms). Additional parameters allow for customization of the plot title and subtitle.
}
\examples{
plotInteractiveHydrograph(Qdat = SampleQ_xts_cfs,
  title= "Colorado River Discharge at Lees Ferry",
  subtitle = "October 3rd 2014 - October 11th 2014")


}
