% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHydrograph.R
\name{plotStaticHydrograph}
\alias{plotStaticHydrograph}
\title{Plot a Static Hydrograph}
\usage{
plotStaticHydrograph(
  Qdat,
  title = "Hydrograph",
  subtitle = "Enter subtitle here",
  unit = "CFS"
)
}
\arguments{
\item{Qdat}{A dataframe containing the discharge data to be plotted.
It must include a POSIXct \code{dateTime} column, and numeric \code{Discharge_cfs}
and \code{Discharge_cms} columns.}

\item{title}{The main title of the hydrograph plot.
Default is "Hydrograph".}

\item{subtitle}{The subtitle of the hydrograph plot, typically used for specifying
the date range or other descriptive information.
Default is "Enter subtitle here".}

\item{unit}{The unit of discharge data to be displayed.
Valid options are 'CFS' for cubic feet per second and 'CMS' for cubic meters per second.
Default is 'CFS'.}
}
\value{
A ggplot object representing the hydrograph.
}
\description{
This function creates a static hydrograph using the ggplot2 package, allowing for
the visualization of water discharge over time in either cubic feet per second (CFS)
or cubic meters per second (CMS). The plot includes options to customize the title and subtitle.
}
\examples{
plotStaticHydrograph(SampleQ_1week_processed, title = "Sample Hydrograph", subtitle = "Sample Data", unit = 'CFS')
}
