% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{printColored}
\alias{printColored}
\title{Print Colored Messages to the Console}
\usage{
printColored(text, color = "green")
}
\arguments{
\item{text}{The message text to be printed.}

\item{color}{A string specifying the color of the text. Supported colors include "green", "blue", and "red".
Default is "green".}
}
\description{
This function prints messages to the R console with specified ANSI colors.
It's designed to help distinguish different types of messages (e.g., success, info, warning)
by color, making them more visually distinguishable.
}
\details{
The function uses ANSI color codes to set the color of the text.
These codes are widely supported in modern terminals and R consoles, including RStudio.
However, some environments, particularly older Windows command lines, might not display
ANSI colors correctly. In such cases, the text will be printed without coloring.
}
\note{
While ANSI color codes are supported in many environments,
there's variability in support and appearance across different platforms and terminals.
Test the function in your target environment to ensure it works as expected.
Additionally, consider providing an option in your functions or scripts to disable colored output
for better compatibility across all environments.
}
\examples{
\dontrun{
printColored("Success message", "green")
printColored("Information message", "blue")
printColored("Error message", "red")
}

}
