% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SampleQ_1week_processed.R,
%   R/data-SampleQ_1week_raw.R
\docType{data}
\name{SampleQ_1week_raw}
\alias{SampleQ_1week_raw}
\title{SampleQ 1 Week processed Data}
\format{
A data frame with 168 observations on the following 3 variables:
\describe{
\item{\code{dateTime}}{A POSIXct date-time indicating the time of the measurement.}
\item{\code{Discharge_cfs}}{Numeric streamflow data, measured in cubic feet per second (cfs).}
\itme{\code}}{Numeric streamflow data, measured in cubic meters per second (cms)}

A data frame with 168 observations on the following 2 variables:
\describe{
\item{\code{dateTime}}{A POSIXct date-time indicating the time of the measurement.}
\item{\code{streamflow}}{Numeric streamflow data, measured in cubic feet per second (cfs).}
}
}
\source{
USGS National Water Information System

USGS National Water Information System
}
\usage{
data(SampleQ_1week_formatted)

data(SampleQ_1week_raw)
}
\description{
This dataset contains raw streamflow data for site 09380000 from the USGS National Water Information System, covering the period from 2014-10-03 to 2014-10-10. The data was retrieved using the \code{dataRetrieval} R package.

This dataset contains raw streamflow data for site 09380000 from the USGS National Water Information System, covering the period from 2014-10-03 to 2014-10-10. The data was retrieved using the \code{dataRetrieval} R package.
}
\keyword{datasets}
