% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SampleQ_xts_cfs.R
\docType{data}
\name{SampleQ_xts_cfs}
\alias{SampleQ_xts_cfs}
\title{SampleQ_xts_cfs: Discharge in cubic feet per second}
\format{
An object of class \code{xts} with 768 rows and 1 column:
\describe{
\item{\code{Discharge_cfs}}{Numeric vector of discharge values in cubic feet per second.}
}
}
\source{
Same as \code{SampleQ_xts_cms}.
}
\usage{
SampleQ_xts_cfs
}
\description{
A dataset containing observations of discharge in cubic feet per second (cfs)
from the same location and period as \code{SampleQ_xts_cms}. Formatted as an xts
object for time series analysis.
}
\keyword{datasets}
