% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SampleQ_xts_cms.R
\docType{data}
\name{SampleQ_xts_cms}
\alias{SampleQ_xts_cms}
\title{SampleQ_xts_cms: Discharge in cubic meters per second}
\format{
An object of class \code{xts} with 768 rows and 1 column:
\describe{
\item{\code{Discharge_cms}}{Numeric vector of discharge values in cubic meters per second.}
}
}
\source{
Same as \code{SampleQ_xts_cfs}.
}
\usage{
SampleQ_xts_cms
}
\description{
A dataset containing observations of discharge in cubic meters per second (cms)
from a specified location over a one-week period in October 2014.
The data is formatted as an xts time series object for easy time series analysis.
}
\details{
This data was created for use in examples for the 'plotInteractiveHydrograph()' function
}
\keyword{datasets}
