% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SandbarSites.R
\docType{data}
\name{SandbarSites}
\alias{SandbarSites}
\title{Sandbar Sites}
\format{
A character vector with 39 unique sandbar site codes.
}
\source{
Describe the source of your data here, if applicable.
}
\usage{
SandbarSites
}
\description{
A character vector containing the identifiers for sandbar sites used in the GrandCanyonSandbaR package.
These sites are referenced by their unique codes and are utilized within the package to fit
generalized power law models to stage-discharge data.
}
\keyword{datasets}
