% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-StageDischargeGrids.R
\docType{data}
\name{StageDischargeGrids}
\alias{StageDischargeGrids}
\title{Stage Discharge Grids}
\format{
A list of data frames where each data frame corresponds to a site and contains
columns:
\itemize{
\item \code{Elevation}: Elevation in meters.
\item \code{CMS_lower}: Lower bound of discharge in cubic meters per second.
\item \code{CMS_median}: Median of discharge in cubic meters per second.
\item \code{CMS_upper}: Upper bound of discharge in cubic meters per second.
\item \code{CFS_lower}: Lower bound of discharge in cubic feet per second.
\item \code{CFS_median}: Median of discharge in cubic feet per second.
\item \code{CFS_upper}: Upper bound of discharge in cubic feet per second.
}
}
\usage{
StageDischargeGrids
}
\description{
A dataset containing the stage discharge relationship modeled
by a Bayesian power law model for multiple sites.
This dataset is used internally by \code{findQ_from_WSE} and \code{findWSE_from_Q} functions.
}
\details{
To access the stage discharge grids for a specific site, use the syntax:
\code{StageDischargeGrids[[sitename]]}, where \code{sitename} is one of the items in the
\code{SandbarSites} dataframe. This dataframe lists all sites for which stage discharge
relationships are available and modeled.
}
\keyword{datasets}
