% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{subsetDatetime}
\alias{subsetDatetime}
\title{Subset discharge data with two timestamps 'YYYYMMDD_HHMM'}
\usage{
subsetDatetime(formattedQ, startDT, endDT)
}
\arguments{
\item{formattedQ}{formatted tibble, output of ConvertQdataRaw2FormattedQ}

\item{startDT}{String. Starting time stamp 'YYYYMMDD_HHMM' ex: '20200101_2315'}

\item{endDT}{String. Ending time stamp 'YYYYMMDD_HHMM' ex: '20220101_2120'}
}
\value{
tibble or data.frame subset by startDT and endDT
}
\description{
Subset discharge data with two timestamps 'YYYYMMDD_HHMM'
}
\examples{
Qdat_Subset<-subsetDatetime(SampleQ_1week_formatted,'20141003_0600', '20140109_1200')
}
