% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageDischargeFunctions.R
\name{validSite}
\alias{validSite}
\title{Validate Site Name}
\usage{
validSite(sitename)
}
\arguments{
\item{sitename}{A character string of the site name to validate.}
}
\value{
None; this function stops execution with an error message if the site name is invalid.
}
\description{
Checks if the provided site name is valid by verifying its presence in the \code{SandbarSites} dataset.
}
\examples{
\dontrun{
  validSite("0034L") # Assumes "0034L" is a valid site in SandbarSites
}
}
