% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHydrograph.R
\name{plotStaticHydrograph}
\alias{plotStaticHydrograph}
\title{Plot a Static Hydrograph}
\usage{
plotStaticHydrograph(
  Qdat,
  title = "Hydrograph",
  subtitle = "Enter subtitle here",
  unit = "CFS"
)
}
\arguments{
\item{Qdat}{A dataframe containing the discharge data to be plotted, expected to include
a POSIXct dateTime column, and numeric Discharge_cfs and Discharge_cms columns.}

\item{title}{The main title of the hydrograph plot. Default is "Hydrograph".}

\item{subtitle}{The subtitle of the hydrograph plot, typically used for specifying
the date range or other descriptive information. Default is "Enter subtitle here".}

\item{unit}{A character string specifying the unit of the discharge data. Supported
units are 'CFS' for cubic feet per second and 'CMS' for cubic meters per second.
Default is 'CFS'.}
}
\value{
A ggplot object representing the hydrograph.
}
\description{
Creates a static hydrograph using the \code{ggplot2} package. This function
is designed to plot time series data of water discharge in either cubic feet per second (CFS)
or cubic meters per second (CMS), with additional parameters to customize the plot title and subtitle.
}
\examples{
plotStaticHydrograph(Qdat = SampleQ_1week_processed,
  title= "Colorado River Discharge at Lees Ferry",
  subtitle = "October 3rd 2014 - October 11th 2014",
  unit = 'CFS')

}
