% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SampleQ_1week_raw.R
\docType{data}
\name{SampleQ_1week_raw}
\alias{SampleQ_1week_raw}
\title{Sample Discharge Data for One Week}
\format{
A data frame with 768 rows and 6 columns:
\describe{
\item{agency_cd}{Character. The agency code. For this dataset, it's always "USGS".}
\item{site_no}{Character. The site number. For this dataset, it's "09380000".}
\item{dateTime}{POSIXct. The date and time of the measurement.}
\item{X_00060_00000}{Double. The discharge measurement in cubic feet per second (cfs).}
\item{X_00060_00000_cd}{Character. The quality code of the measurement. "A" indicates approved data.}
\item{tz_cd}{Character. The time zone code. For this dataset, it's always "UTC".}
}
}
\source{
USGS Water Data for the Nation
}
\usage{
SampleQ_1week_raw
}
\description{
A dataset containing raw discharge measurements for a specific USGS site
over a one-week period from October 3, 2014, to October 10, 2014. Data was
retrieved using the \code{\link[dataRetrieval]{readNWISuv}} function from
the \pkg{dataRetrieval} package.
}
\references{
\url{https://waterdata.usgs.gov/nwis}
}
\keyword{datasets}
